+++ cli/src/main.rs
@@ -1,21 +1,30 @@
+use std::fs;
 use std::io::{self, Read};
 use std::path::Path;
 
 use atty::Stream;
 use bonsol_sdk::BonsolClient;
 use clap::Parser;
+use common::{execute_get_inputs, ZkProgramManifest};
+use risc0_circuit_rv32im::prove::emu::exec::DEFAULT_SEGMENT_LIMIT_PO2;
+use risc0_circuit_rv32im::prove::emu::testutil::DEFAULT_SESSION_LIMIT;
+use risc0_zkvm::ExecutorEnv;
 use solana_sdk::signature::read_keypair_file;
 use solana_sdk::signer::Signer;
 
 use crate::command::{BonsolCli, ParsedBonsolCli, ParsedCommand};
 use crate::common::{sol_check, try_load_from_config};
-use crate::error::BonsolCliError;
+use crate::error::{BonsolCliError, ZkManifestError};
 
 mod build;
 mod deploy;
+mod estimate;
 mod execute;
 mod init;
 mod prove;
+
+#[cfg(all(test, feature = "integration"))]
+mod tests;
 
 pub mod command;
 pub mod common;
@@ -58,6 +67,42 @@ async fn main() -> anyhow::Result<()> {
                 .into());
             }
             deploy::deploy(rpc, keypair, deploy_args).await
+        }
+        ParsedCommand::Estimate {
+            manifest_path,
+            input_file,
+            max_cycles,
+        } => {
+            let manifest_file = fs::File::open(Path::new(&manifest_path)).map_err(|err| {
+                BonsolCliError::ZkManifestError(ZkManifestError::FailedToOpen {
+                    manifest_path: manifest_path.clone(),
+                    err,
+                })
+            })?;
+            let manifest: ZkProgramManifest =
+                serde_json::from_reader(manifest_file).map_err(|err| {
+                    BonsolCliError::ZkManifestError(ZkManifestError::FailedDeserialization {
+                        manifest_path,
+                        err,
+                    })
+                })?;
+            let elf = fs::read(&manifest.binary_path).map_err(|err| {
+                BonsolCliError::ZkManifestError(ZkManifestError::FailedToLoadBinary {
+                    binary_path: manifest.binary_path.clone(),
+                    err,
+                })
+            })?;
+            let mut env = &mut ExecutorEnv::builder();
+            env = env
+                .segment_limit_po2(DEFAULT_SEGMENT_LIMIT_PO2 as u32)
+                .session_limit(max_cycles.or(DEFAULT_SESSION_LIMIT));
+
+            if input_file.is_some() {
+                let inputs = execute_get_inputs(input_file, None)?;
+                let inputs: Vec<&str> = inputs.iter().map(|i| i.data.as_str()).collect();
+                env = env.write(&inputs.as_slice())?;
+            }
+            estimate::estimate(elf.as_slice(), env.build()?)
         }
         ParsedCommand::Execute {
             execution_request_file,